api-server/requirements.txt
Django~=5.1.2
django-cors-headers~=4.9.0
djangorestframework~=3.15.2

mysqlclient>=2.2.0
python-dotenv>=1.0.0

django-request-id>=1.0.0